requirements.txt
djangorestframework
django-filter==23.2
